testHackBangkok/requirements.txt
flask==3.0.2
openai>=1.12.0
pandas==2.2.1
python-dotenv==1.0.1
openpyxl==3.1.2 